move home_screen.dart to /lib/home
