+++ lib/utils/routing/app_router.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_boilerplate_hng11/features/auth/screen/splash_screen.dart';
 import 'package:flutter_boilerplate_hng11/features/cart/cart_screen.dart';
-import 'package:flutter_boilerplate_hng11/features/main_view/home_screen.dart';
+import 'package:flutter_boilerplate_hng11/features/home/home_screen.dart';
 import 'package:flutter_boilerplate_hng11/features/main_view/main_view.dart';
 import 'package:flutter_boilerplate_hng11/features/product_listing/screens/product_screen.dart';
 import 'package:flutter_boilerplate_hng11/features/user_setting/screens/account_settings.dart';
